publs/pyproject.toml
[project]
name = "publs"
version = "0.2.0"
description = "Maintain a single SSOT BibTeX file for SLDS publications, with interactive suggestions from OpenAlex / Crossref / Google Scholar."
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0",
    "click>=8.1",
    "bibtexparser>=1.4",
    "httpx>=0.27",
]

[project.scripts]
publs = "publs.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/publs"]
